fix: make applicationData test robust by sharing server lifecycle
Each of the 8 tests was starting and stopping a full server with 3
login requests, causing intermittent timeouts under load. Restructure
to start the server once in before() and clean applicationData in
beforeEach(). Convert to TypeScript and fix the 'foo/bar' test case
to use encodeURIComponent so it actually tests invalid appid validation.
